+++ docs/changelog.md
@@ -21,6 +21,7 @@ PySceneDetect Releases
  - [bugfix] Fix `scenedetect.detect()` throwing `TypeError` when specifying `stats_file_path`
  - [enhancement] Add optional `start_time` and `end_time` arguments to `scenedetect.detect()`
  - [feature] Add new backend `VideoStreamMoviePy` using the MoviePy package
+ - [enhancement] If available, the `ffmpeg` binary from the `imageio_ffmpeg` package will be used if one could not be found in PATH
 
 ### 0.6 (May 29, 2022)